i18n via Routing
This commits adds fr/ and en/ routes; also uses browser language to initialize the language in store and then changes the URL; next the URL is checked to see if the language is there or else using browser language.
A lot of boilerplate duplicate code to handle this, I'm not quite satisfied by this solution so I will create another branch with another solution.
Because all the pages are redirected and need to be duplicated in the fr and en routes ; but I want to keep the old non-i18n routes for SEO and old links.

Tests are not fixed.